gcd.py works

+++ misc/gcd.py
@@ -4,5 +4,4 @@ import sys
 gcd = lambda x, y : x if y == 0 else gcd(y, x%y)
 
 assert( len(sys.argv) == 3 )
-print( gcd(sys.argv[1], sys.argv[2]) )
-
+print( gcd(long(sys.argv[1]), long(sys.argv[2])) )
